+++ src/SerializerGenerator.php
@@ -240,7 +240,7 @@ final class SerializerGenerator
                 $innerCode = $this->generateCodeForClass($subType->getClassMetadata(), $apiVersion, $serializerGroups, $arrayPath.'['.$index.']', $modelPath.'['.$index.']', $stack);
                 break;
 
-            case $subType instanceof PropertyTypeUnknown && $this->configuration->shouldAssignUnknownArrays():
+            case $subType instanceof PropertyTypeUnknown && $this->configuration->shouldAllowGenericArrays():
                 return $this->templating->renderArrayAssign($arrayPath, $modelPath);
 
             default: